requirements.txt
Flask==0.10.1
Flask-Script==0.6.7
Flask-WTF==0.9.4
Flask-SQLAlchemy==1.0
flask-debugtoolbar==0.9.0
Flask-Uploads==0.1.3
Flask-Mail==0.9.0
#MySQL-python==1.2.5
gunicorn==18.0
alioss==0.22.4
fabric==1.8.3
Pillow==2.4.0
Flask-Migrate
redis
Celery
requests
blinker
raven

# test
nose
coverage
pylint

python-daemon
gevent
qrcode